+++ src/main/java/com/webank/webase/front/trade/request/WithDrawReq.java
@@ -2,6 +2,7 @@ package com.webank.webase.front.trade.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Data
@@ -9,7 +10,7 @@ public class WithDrawReq {
 
     private String contractId;
     private String secerte;
-    private BigInteger value;
+    private BigDecimal value;
     private String  partnerAssetAddress;
     private BigInteger  partnerAssetMinunit;
